refactor: State book 패키지 -> state 패키지로 이동 및 BookState 이넘값 끝에 STATE 키워드 추가
