+++ src/main/java/library/book/domain/constants/BookState.java
@@ -2,7 +2,7 @@ package library.book.domain.constants;
 
 import java.util.function.Supplier;
 
-import library.book.domain.State;
+import library.book.domain.state.State;
 import library.book.domain.state.AvailableRent;
 import library.book.domain.state.Cleaning;
 import library.book.domain.state.Lost;
@@ -10,10 +10,10 @@ import library.book.domain.state.Rented;
 
 public enum BookState {
 
-	AVAILABLE_RENT("대여 가능", AvailableRent::new),
-	RENTED("대여중", Rented::new),
-	CLEANING("정리중", Cleaning::new),
-	LOST("분실", Lost::new);
+	AVAILABLE_RENT_STATE("대여 가능", AvailableRent::new),
+	RENTED_STATE("대여중", Rented::new),
+	CLEANING_STATE("정리중", Cleaning::new),
+	LOST_STATE("분실", Lost::new);
 
 	private final String description;
 	private final Supplier<State> stateSupplier;
