+++ src/main/java/library/book/domain/state/AvailableRent.java
@@ -3,13 +3,12 @@ package library.book.domain.state;
 import static library.book.domain.constants.BookState.*;
 import static library.book.exception.ErrorCode.*;
 
-import library.book.domain.State;
 import library.book.domain.constants.BookState;
 import library.book.exception.BookException;
 
 public class AvailableRent implements State {
 
-	private final BookState bookState = AVAILABLE_RENT;
+	private final BookState bookState = AVAILABLE_RENT_STATE;
 
 	@Override
 	public BookState getBookState() {
