+++ src/main/java/library/book/domain/state/Cleaning.java
@@ -3,7 +3,6 @@ package library.book.domain.state;
 import static library.book.domain.constants.BookState.*;
 import static library.book.exception.ErrorCode.*;
 
-import library.book.domain.State;
 import library.book.domain.constants.BookState;
 import library.book.exception.BookException;
 
@@ -13,7 +12,7 @@ public class Cleaning implements State {
 
 	@Override
 	public BookState getBookState() {
-		return CLEANING;
+		return CLEANING_STATE;
 	}
 
 	@Override
