+++ src/test/java/library/book/domain/StateTest.java
@@ -12,6 +12,7 @@ import library.book.domain.state.AvailableRent;
 import library.book.domain.state.Cleaning;
 import library.book.domain.state.Lost;
 import library.book.domain.state.Rented;
+import library.book.domain.state.State;
 import library.book.exception.BookException;
 
 @DisplayName("[State 테스트] - Domain")
